+++ HISTORY.rst
@@ -3,6 +3,7 @@ History
 
 Next Release
 ------------
+* Update docstring of test_find_duplicate_reactions.
 
 0.9.6 (2019-03-06)
 ------------------