+++ HISTORY.rst
@@ -4,6 +4,7 @@ History
 Next Release
 ------------
 * Update docstring of test_find_duplicate_reactions.
+* Add guidelines on writing custom tests to the documentation.
 
 0.9.6 (2019-03-06)
 ------------------